% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunSEMA.R
\name{AddSEMAtoSeurat}
\alias{AddSEMAtoSeurat}
\title{Add SEMA results to Seurat object}
\usage{
AddSEMAtoSeurat(
  seurat_obj,
  sema_result,
  prefix = "SEMA_",
  use_spatial_scores = TRUE
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{sema_result}{SEMA result object}

\item{prefix}{Prefix to add to metadata column names, default "SEMA_"}

\item{use_spatial_scores}{Whether to use spatial scores (TRUE) or raw scores (FALSE), default TRUE}
}
\value{
Updated Seurat object
}
\description{
Add SEMA-calculated spatial scores to Seurat object metadata
}
