% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateData.R
\name{build_coef_data}
\alias{build_coef_data}
\title{build_coef_data creates a data set with random coefficients per unit.}
\usage{
build_coef_data(j, random_coef_sd)
}
\arguments{
\item{j}{Number of units.}

\item{random_coef_sd}{A vector with true variance of the random effects.}
}
\value{
A data frame random coefficients for all random effects for
  all units.
}
\description{
build_coef_data creates a data set with random coefficients per unit.
}
