% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateData.R
\name{build_dataset}
\alias{build_dataset}
\title{Generate a multilevel long-form data frame}
\usage{
build_dataset(n, j, fixed_coef, random_coef_sd, resid_sd, n_level_2_var = 2,
  mean_fixed_level_2 = 0, sd_fixed_level_2 = 1, mean_fixed_level_1 = 0,
  sd_fixed_level_1 = 1)
}
\arguments{
\item{n}{Number of observations.}

\item{j}{Number of individuals.}

\item{fixed_coef}{Vector with fixed effects coefficients.}

\item{random_coef_sd}{Vector with standard deviations of the random effects.}

\item{resid_sd}{Scalar, residual variance.}

\item{n_level_2_var}{Number of level 2 variables.}

\item{mean_fixed_level_2}{Means of the fixed effects covariates, level 2.}

\item{sd_fixed_level_2}{Standard Deviation of the fixed effects covariates,
level 2.}

\item{mean_fixed_level_1}{Means of the fixed effects covariates, level 1.}

\item{sd_fixed_level_1}{Standard deviation of the fixed effects covariates,
level 1.}
}
\value{
A data frame with variable \code{id}, which labels the units,
  \code{y} is the outcome or dependent variable and covariates.
}
\description{
This function builds a long-form data set with \code{n} 
observations from \code{j} units.
}
\details{
The function creates a data frame containing variables \code{id}, 
  \code{y}, and \code{V#} with the number of fixed effects. The user can 
  specify the number of level 2 covariates. The data are generated in two 
  steps, first the level 2 data are generated, which are constant per unit
  then the ordering of units is created using the \code{sample} function, 
  with \code{replace = TRUE}. This results in unequal number of observations 
  per unit, altough each unit has equal probabilty of being sampled. Lastly, 
  level 1 data are generated by drawing from the normal distribution.
}
\keyword{dataset}
\keyword{multilevel}
