% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateData.R
\name{build_person_data}
\alias{build_person_data}
\title{build_person_data is a function to create level 2 data.}
\usage{
build_person_data(j, n_level_2_var, mean_fixed_level_2, sd_fixed_level_2)
}
\arguments{
\item{j}{Number of units.}

\item{n_level_2_var}{Number of level 2 variables.}

\item{mean_fixed_level_2}{A vector with means for the level 2 covariates.}

\item{sd_fixed_level_2}{A vector with standard deviations for the level 2
covariates.}
}
\value{
A data frame with level 2 data for the units.
}
\description{
build_person_data is a function to create level 2 data.
}
