% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{CalculateMultipleSEMAs}
\alias{CalculateMultipleSEMAs}
\title{Calculate multiple gene sets in batch}
\usage{
CalculateMultipleSEMAs(
  expression_matrix,
  gene_sets,
  coords,
  k = 30,
  raw_method = "mean",
  spatial_weight = 0.5,
  bandwidth = NULL,
  addmodulescore_params = list(),
  verbose = TRUE,
  n_cores = 1
)
}
\arguments{
\item{expression_matrix}{Expression matrix}

\item{gene_sets}{Gene set list}

\item{coords}{Coordinate matrix}

\item{k}{Number of nearest neighbors, default 30}

\item{raw_method}{Raw score calculation method}

\item{spatial_weight}{Spatial smoothing weight}

\item{bandwidth}{Gaussian kernel bandwidth}

\item{addmodulescore_params}{AddModuleScore parameters}

\item{verbose}{Whether to display detailed information}

\item{n_cores}{Number of parallel cores, default 1 (serial)}
}
\value{
List containing all gene set results
}
\description{
Calculate spatial scores for multiple gene sets simultaneously
}
