% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estep.R
\name{compute_c_inv}
\alias{compute_c_inv}
\title{compute_c_inv computes the uncertainty of the random effects coefficients.
  NOTE, this function contains a matrix inversion of the variance of the
  random effects, when this matrix is large, the SEMA algorithm slows down
  from raudenbush and bryk (2002) Hierarchial linear models, 2nd edition,
  EQUATION 14.15b}
\usage{
compute_c_inv(z_sq, resid_var, random_var)
}
\arguments{
\item{z_sq}{Square matrix of the squared random effects variables.}

\item{resid_var}{Residual variance, a scalar.}

\item{random_var}{Variance of the random effects, a matrix.}
}
\value{
Inverted matrix with updated uncertainty of the random effects
  coefficients.
}
\description{
compute_c_inv computes the uncertainty of the random effects coefficients.
  NOTE, this function contains a matrix inversion of the variance of the
  random effects, when this matrix is large, the SEMA algorithm slows down
  from raudenbush and bryk (2002) Hierarchial linear models, 2nd edition,
  EQUATION 14.15b
}
