% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mstep.R
\name{compute_fixed_coef}
\alias{compute_fixed_coef}
\title{compute_fixed_coef computes the coefficients of the fixed effects,
  see from raudenbush and bryk (2002) Hierarchial linear models, 2nd edition,
  EQUATION 14.10.}
\usage{
compute_fixed_coef(x_inv, xy, t1)
}
\arguments{
\item{x_inv}{The inverted matrix of the fixed effects variables.}

\item{xy}{The product of fixed effects variables and dependent variable.}

\item{t1}{A vector with Complete Data Sufficient Statistics of the
fixed effects.}
}
\value{
A vector with fixed effects coefficients.
}
\description{
compute_fixed_coef computes the coefficients of the fixed effects,
  see from raudenbush and bryk (2002) Hierarchial linear models, 2nd edition,
  EQUATION 14.10.
}
