% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estep.R
\name{compute_random_coef}
\alias{compute_random_coef}
\title{compute_random_coef computes the random effects: intercepts and slopes
  from raudenbush and bryk (2002) Hierarchial linear models, 2nd edition,
  EQUATION 14.15a}
\usage{
compute_random_coef(c_inv, fixed_coef, zy, zx)
}
\arguments{
\item{c_inv}{The inverse of the uncertainty matrix,}

\item{fixed_coef}{A vector with coefficients of the fixed effects.}

\item{zy}{The product of random effects variables and dependent variable,
a vector.}

\item{zx}{The product of random effects variables and fixed effect 
variables, a matrix.}
}
\value{
A vector with predicted random effects coefficients.
}
\description{
compute_random_coef computes the random effects: intercepts and slopes
  from raudenbush and bryk (2002) Hierarchial linear models, 2nd edition,
  EQUATION 14.15a
}
\seealso{
\code{\link{compute_c_inv}}.
}
