% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mstep.R
\name{compute_resid_var}
\alias{compute_resid_var}
\title{compute_resid_var computes the residual variance.}
\usage{
compute_resid_var(t3, n)
}
\arguments{
\item{t3}{The Complete Data Sufficient Statistic of the residual variance,
a scalar.}

\item{n}{The number of observations.}
}
\value{
The residual variance (scalar).
}
\description{
compute_resid_var computes the residual variance.
}
