% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estep.R
\name{compute_t1_j}
\alias{compute_t1_j}
\title{compute_t1_j computes the individual contribution to the
  complete data sufficient statistics of the fixed effects.}
\usage{
compute_t1_j(zx, random_coef)
}
\arguments{
\item{zx}{The product of the random effects and fixed effects variables,
a matrix.}

\item{random_coef}{A vector of random effects coefficients,}
}
\value{
A vector of individual contributions to the CDSS of the fixed
  effects.
}
\description{
compute_t1_j computes the individual contribution to the
  complete data sufficient statistics of the fixed effects.
}
\seealso{
\code{\link{compute_random_coef}}.
}
