% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estep.R
\name{compute_t2_j}
\alias{compute_t2_j}
\title{compute_t2_j computes the individual contribution to the complete data
  sufficient statistics of the random effects.}
\usage{
compute_t2_j(random_coef, resid_var, c_inv)
}
\arguments{
\item{random_coef}{A vector of random effects coefficients,
@seealso  \code{\link{compute_random_coef}}.}

\item{resid_var}{The residual variance, a scalar.}

\item{c_inv}{The uncertainty matrix of the random effects coefficients,
@seealso  \code{\link{compute_random_coef}}.}
}
\value{
A matrix of individual contributions to the CDSS of variance the
  random effects.
}
\description{
compute_t2_j computes the individual contribution to the complete data
  sufficient statistics of the random effects.
}
