% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estep.R
\name{compute_t3_j}
\alias{compute_t3_j}
\title{compute_t3_j computes the individual contribution to the complete data
  sufficient statistics of the residual variance.}
\usage{
compute_t3_j(parameters_j, random_coef, fixed_coef, resid_var)
}
\arguments{
\item{parameters_j}{A list with unit parameters:
y_sq, a scalar of #' the sum of the squared responses; x_sq,
a matrix with outer product of the fixed effects covariates;
z_sq, a matrix with outer product of the random effects covariates;
xy, a vector with the product of fixed effects covariates and the response
variable;
zy, a vector with the product of random effects covariates and the response
variable;
xz, product of fixed effects covariates and random effects covariates,
a matrix;
c_inv, the inverse of the uncertainty matrix of the
random effects coefficients,
@seealso \code{\link{compute_c_inv}}.}

\item{random_coef}{random effects coefficients,
@seealso  \code{\link{compute_random_coef}}.}

\item{fixed_coef}{A vector of fixed effects coefficients.}

\item{resid_var}{The residual variance, a scalar.}
}
\value{
A scalar with unit's contribution to the Complete Data
  Sufficient Statistics of the residual variance.
}
\description{
compute_t3_j computes the individual contribution to the complete data
  sufficient statistics of the residual variance.
}
