% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitiateGlobalParameters.R
\name{create_theta_main}
\alias{create_theta_main}
\title{Sema model list}
\usage{
create_theta_main(n_fixed, n_random, start_resid_var = 1,
  start_random_var = 1, start_fixed_coef = NULL, prior_n, prior_j)
}
\arguments{
\item{n_fixed}{Number of fixed effects.}

\item{n_random}{Number of random effects.}

\item{start_resid_var}{Start values residual variance, default start
value equals 1.}

\item{start_random_var}{Start values variance of random effects,
default start values equal 1. Make sure that the length of 
\code{start_random_var} matches the number of random effects.}

\item{start_fixed_coef}{The default is set to NULL, when no start values 
are provided, \code{start_fixed_coef} is set to \code{rep(1, n_fixed)} 
within the \code{create_theta_main} function. Note that the length of the
start values must match the number of fixed effects.}

\item{prior_n}{A scalar indicating the weight of the start value of the 
residual variance.}

\item{prior_j}{A scalar indicating the weight of the start values of the
fixed effects coefficients and random effects variance.}
}
\value{
A list which contains: the fixed effects coefficients, 
   \code{fixed_coef_hat}; the Complete Data Sufficient Statistic of the 
   fixed effects, \code{t1}; the variance of the random effects 
   \code{random_var_hat}; the Complete Data Sufficent Statistics of the 
   variance of random effects, \code{t2}; the residual variance, 
   \code{resid_var_hat}; the Complete Data Sufficiant Statistic of
   the residual variance, \code{t3}, the number of observations, \code{n};
   the number of individuals, \code{j}; the average dependent variable, 
   which can be used for prediction purposes but it is not needed for the
   algorithm, \code{y}; the square matrix of squared fixed effects, which is
   only updated as long as the matrix is not yet invertible, once inverted,
   only the inverted matrix is updated, \code{x_sq}; the product of fixed 
   effects and dependent variable, \code{xy_vector}; the inverse of the 
   \code{x_sq} matrix, \code{x_inv}.
}
\description{
This function creates a list with all required objects at the global level
  needed by the sema algorithm
}
\details{
The function creates a list with objects used by the sema 
  algorithm, most are not used for interpretation. The user can provide 
  start values to enhance the algorithm's performance. In that case, the
  function also requires the weight of those start values, 
  when no starting values are given default values are used.
}
\keyword{fitting}
\keyword{method}
\keyword{model}
\keyword{sema}
\keyword{start}
