% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{EstimateBandwidth}
\alias{EstimateBandwidth}
\title{Estimate Gaussian kernel bandwidth}
\usage{
EstimateBandwidth(coords, method = "median_distance", k = 10)
}
\arguments{
\item{coords}{Coordinate matrix}

\item{method}{Estimation method: "median_distance" (default) or "mean_distance"}

\item{k}{Number of nearest neighbors for estimation, default 10}
}
\value{
Estimated bandwidth value
}
\description{
Automatically estimate bandwidth parameter for Gaussian kernel
}
