% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{fixef}
\alias{fixef}
\title{Extract the fixed effects coefficients.}
\usage{
fixef(x)
}
\arguments{
\item{x}{The sema output, a list with unit parameters.}
}
\value{
A data frame with fixed effect coefficients.
}
\description{
This function returns the estimated fixed coefficients.
}
\keyword{coefficients}
\keyword{model}
