% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEMAPredict.R
\name{get_theta}
\alias{get_theta}
\title{get model statistics from Sema object}
\usage{
get_theta(model)
}
\arguments{
\item{model}{sema object}
}
\description{
This function returns a list with all model statistics. The
  function is for instance used in the predict function.
}
