% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEMAPredict.R
\name{get_theta_j}
\alias{get_theta_j}
\title{get unit statistics from Sema object}
\usage{
get_theta_j(model, id)
}
\arguments{
\item{model}{sema object}

\item{id}{unit's identifier}
}
\description{
This function returns a list with all unit statistics. The
  function is for instance used in the predict function.
}
