% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{is_sema}
\alias{is_sema}
\title{sema output}
\usage{
is_sema(x)
}
\arguments{
\item{x}{Sema output which should be a list.}
}
\value{
LOGICAL
}
\description{
\code{is_sema} checks whether the input is sema output
}
