% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estep.R
\name{online_e_step}
\alias{online_e_step}
\title{online_e_step includes the latter three functions, for short hand notation
  this function retrieves all the needed information from the global storage
  and individual storage.
  See from raudenbush and bryk (2002) Hierarchial linear models, 2nd edition,
  EQUATION 14.23}
\usage{
online_e_step(theta_main, theta_id)
}
\arguments{
\item{theta_main}{A list current state of model parameters.}

\item{theta_id}{A list with information of the current entering individual.}
}
\value{
A list with contributions and new CDSS,
  these do still need to be split and returned to the right objects.
}
\description{
online_e_step includes the latter three functions, for short hand notation
  this function retrieves all the needed information from the global storage
  and individual storage.
  See from raudenbush and bryk (2002) Hierarchial linear models, 2nd edition,
  EQUATION 14.23
}
