% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEMAPredict.R
\name{predict}
\alias{predict}
\title{SEMA predict function}
\usage{
predict(theta = get_theta(), theta_j = get_theta_j(), data_fixed,
  data_random)
}
\arguments{
\item{theta}{The current state of model parameters.}

\item{theta_j}{The current state of the individual parameters.}

\item{data_fixed}{The current observation fixed effects variables.}

\item{data_random}{The current observation random effects variables.}
}
\description{
This function returns the predicted scores based on the current 
  parameter estimates and the data currently entering of this unit.
}
\examples{
## First we create a dataset, consisting of 2500 observations from 20 
## units. The fixed effects have the coefficients 1, 2, 3, 4, and 5. The 
## variance of the random effects equals 1, 4, and 9. Lastly the 
## residual variance equals 4:
test_data <- build_dataset(n = 1500, 
                           j = 200, 
                           fixed_coef = 1:5, 
                           random_coef_sd = 1:3, 
                           resid_sd = 2)
                           
## fit a multilevel model: 
m1 <- sema_fit_df(formula = y ~ 1 + V3 + V4 + V5 + V6 + (1 + V4 + V5  | id), 
                   data_frame = test_data, intercept = TRUE)
## predict for the last row of the dataset:                     
predict_unit_outcome <- predict(theta = get_theta(model = m1), 
                               theta_j = get_theta_j(model = m1, 
                                 id = test_data$id[2000]),
                               data_fixed = as.numeric(
                                 test_data[2000, 3:7]),
                               data_random = as.numeric(
                                 test_data[2000, c(3, 5, 6)]))
}
\keyword{prediction}
