% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{print.SEMAresult}
\alias{print.SEMAresult}
\title{Print SEMA result summary}
\usage{
\method{print}{SEMAresult}(x, ...)
}
\arguments{
\item{x}{SEMA result object}

\item{...}{Other parameters}
}
\description{
Print SEMA result summary
}
