% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{ranef}
\alias{ranef}
\title{Extract the random effects coefficients.}
\usage{
ranef(x)
}
\arguments{
\item{x}{The sema output, a list with unit parameters.}
}
\value{
A data frame with random effect coefficients per unit, sorted by
  entry.
}
\description{
This function loops over the list with unit parameters and 
  returns a dataframe with the random coefficients, centered around zero.
}
\details{
The \code{ranef} function works similar to lme4's 
  \code{ranef}. The function uses \code{ldply} function from the 
  \code{plyr} package to extract all coefficients.
}
\examples{
## First we create a dataset, consisting of 2500 observations from 20 
## units. The fixed effects have the coefficients 1, 2, 3, 4, and 5. The 
## variance of the random effects equals 1, 4, and 9. Lastly the 
## residual variance equals 4:
  
test_data <- build_dataset(n = 2500, 
                           j = 20, 
                           fixed_coef = 1:5, 
                           random_coef_sd = 1:3, 
                           resid_sd = 2)
                           
## Next, we fit a simple model to these data                           
m1 <- sema_fit_df(formula = y ~ 1 + V3 + V4 + V5 + V6 + (1 + V4 + V5 | id), 
                    data_frame = test_data, 
                    intercept = TRUE)
ranef(m1)
}
\keyword{coefficients}
\keyword{model}
