% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEMA.R
\docType{package}
\name{SEMA-package}
\alias{SEMA}
\alias{SEMA-package}
\title{Streaming EM Approximation algorithm---multilevel model: fixed and random 
  effects}
\description{
The \code{SEMA} package allows one to fit online, or in a data stream, 
  multilevel models including fixed and random intercepts and slopes. 
  The equation numbers refer to the equations in Raudenbush and Bryk 
  hierarchical linear models, 2nd edition.
}
\section{The \code{SEMA} package includes the following functions}{

\itemize{
  \item \code{\link{build_dataset}}: creates a multilevel, long-form, 
    data set.  
  \item \code{\link{build_theta_j}}: creates all objects at the unit level.
  \item \code{\link{create_theta_main}}: allows users to include own start 
    values.
  \item \code{\link{is.sema}}: checks whether list is sema class.
  \item \code{\link{sema_fit_df}}: Fits multilevel models, input a 
    data.frame, returns model parameters, units' parameters, and optional, 
    stores the model parameters during the updating process.   
  \item \code{\link{sema_fit_one}}: Fits multilevel models, input one data
    vector, returns updated model and unit parameters.
  \item \code{\link{sema_fit_set}}: Fits multilevel models, input one data
    vector, returns updated model parameters and a list with all units' 
    parameters, one of them updated. 
  \item \code{\link{predict}}: predicts outcome based on the data of 
    the covariates.
  \item \code{\link{ranef}}: returns a data frame with coefficients of
    the random effects.
  \item \code{\link{store_fixed_coef}}: stores the estimates of the 
    fixed effects coefficients.
  \item \code{\link{store_random_var}}: stores the estimates of the 
    variance of the random effects. 
  \item \code{\link{store_resid_var}}: stores the estimate of residual 
    variance coefficients 
  \item \code{\link{summary_sema}}: returns a list with the number of 
    units, number of observations, estimates of the fixed effects 
    coefficients, estimates of the variance of the random effects and the 
    estimate of the residual variance 
}
}

\author{
\strong{Maintainer}: Lianne Ippel \email{g.j.e.ippel@gmail.com}

}
