% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEMAPredict.R
\name{sema_predict}
\alias{sema_predict}
\title{SEMA predict function}
\usage{
sema_predict(theta, theta_j, data_fixed, data_random)
}
\arguments{
\item{theta}{The current state of model parameters.}

\item{theta_j}{The current state of the individual parameters.}

\item{data_fixed}{The current observation fixed effects variables.}

\item{data_random}{The current observation random effects variables.}
}
\description{
This function returns the predicted scores based on the current 
  parameter estimates and the data currently entering of this unit.
}
\keyword{prediction}
