% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{sema_ranef}
\alias{sema_ranef}
\title{Extract the random effects coefficients.}
\usage{
sema_ranef(x)
}
\arguments{
\item{x}{The sema output, a list with unit parameters.}
}
\value{
A data frame with random effect coefficients per unit, sorted by
  entry.
}
\description{
This function loops over the list with unit parameters and 
  returns a dataframe with the random coefficients, centered around zero.
}
\details{
The \code{sema_ranef} function works similar to lme4's 
  \code{ranef}. The function uses \code{ldply} function from the 
  \code{plyr} package to extract all coefficients.
}
\keyword{coefficients}
\keyword{model}
