% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{sema_store_random_var}
\alias{sema_store_random_var}
\title{Store the variance of the random effects}
\usage{
sema_store_random_var(sema_model)
}
\arguments{
\item{sema_model}{The sema output, a list containing the model parameters.}
}
\value{
A data frame with the number of observations and the random effects
  variance.
}
\description{
This function extracts the variance of the random effects and 
saves them in a data frame with the number of data points seen so far.
}
\seealso{
\code{\link{sema_ranef}}, \code{\link{sema_store_fixed_coef}}, 
  \code{\link{sema_store_resid_var}}
}
\keyword{coefficients}
\keyword{save}
