% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{sema_store_resid_var}
\alias{sema_store_resid_var}
\title{store residual variance}
\usage{
sema_store_resid_var(sema_model)
}
\arguments{
\item{sema_model}{The sema output, a list containing the model parameters.}
}
\value{
A data frame with the number of observations and the residual
  variance.
}
\description{
This function extracts the residual variance and saves
  it in a data frame with the number of data points seen so far.
}
\seealso{
\code{\link{sema_ranef}}, \code{\link{sema_store_fixed_coef}},
 \code{\link{sema_store_random_var}}
}
\keyword{coefficients}
\keyword{save}
