% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{store_random_var}
\alias{store_random_var}
\title{Store the variance of the random effects}
\usage{
store_random_var(object)
}
\arguments{
\item{object}{The sema output, a list containing the model parameters.}
}
\value{
A data frame with the number of observations and the random effects
  variance.
}
\description{
This function extracts the variance of the random effects and 
saves them in a data frame with the number of data points seen so far.
}
\seealso{
\code{\link{ranef}}, \code{\link{fixef}}, 
  \code{\link{store_fixed_coef}}, 
  \code{\link{store_resid_var}}
}
\keyword{coefficients}
\keyword{save}
