% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{store_resid_var}
\alias{store_resid_var}
\title{store residual variance}
\usage{
store_resid_var(object)
}
\arguments{
\item{object}{The sema output, a list containing the model parameters.}
}
\value{
A data frame with the number of observations and the residual
  variance.
}
\description{
This function extracts the residual variance and saves
  it in a data frame with the number of data points seen so far.
}
\seealso{
\code{\link{ranef}}, \code{\link{store_fixed_coef}},
 \code{\link{store_random_var}}
}
\keyword{coefficients}
\keyword{save}
