% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_sema}
\alias{summary_sema}
\title{Interpreting sema output}
\usage{
summary_sema(x)
}
\arguments{
\item{x}{A sema model output.}
}
\value{
A list with sample size, number of units, the coefficients of the
  fixed effects, the variance of the random effects and the residual
  variance.
}
\description{
Returns a list with the current model parameter estimates
}
\details{
The output of the sema_fit functions are usually large, and 
  difficult to read, lists. In order to interpret the output of the sema_fit
  functions, \code{summary_sema} returns a general overview of the model 
  parameters, the fixed effects coefficients, the random effects variances
  and covariances, and the residual variance. The \code{summary_sema} 
  function is also incorpprated within the \code{sema_fit_one} function, 
  such that with the argument \code{print_every} the user can see a summary
  of the updated model parameters every X data points.
}
\examples{
## First we create a dataset, consisting of 2500 observations from 20 
## units. The fixed effects have the coefficients 1, 2, 3, 4, and 5. The 
## variance of the random effects equals 1, 4, and 9. Lastly the 
## residual variance equals 4:
  
test_data <- build_dataset(n = 2500, 
                           j = 20, 
                           fixed_coef = 1:5, 
                           random_coef_sd = 1:3, 
                           resid_sd = 2)
                           
## Next, we fit a simple model to these data                           
m1 <- sema_fit_df(formula = y ~ 1 + V3 + V4 + V5 + V6 + (1 + V4 + V5 | id), 
                    data_frame = test_data, 
                    intercept = TRUE)
summary_sema(m1)
}
\seealso{
\code{\link{store_fixed_coef}}, 
   \code{\link{store_random_var}}, \code{\link{store_resid_var}}, 
   \code{\link{ranef}}
}
\keyword{model}
\keyword{sema}
\keyword{summary}
