% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mstep.R
\name{try_solve}
\alias{try_solve}
\title{Test whether matrix is invertible}
\usage{
try_solve(x_sq)
}
\arguments{
\item{x_sq}{A square matrix of fixed effects variables squared.}
}
\value{
If succesful: the inverted matrix. If not succesful: NULL
}
\description{
\code{try_solve} is a function which tries to invert the 
  squared fixed effects covariates matrix.
}
