% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitiateGlobalParameters.R
\name{try_theta}
\alias{try_theta}
\title{try_theta is a function which creates the model parameters and objects
  related to them. If the input is not a list, this function creates a list
  which is also a "sema" class with all needed objects stored within.}
\usage{
try_theta(theta, n_fixed, n_random, start_resid_var, start_random_var,
  start_fixed_coef, prior_n, prior_j)
}
\arguments{
\item{theta}{Either an empty object or an object with class list and sema.}

\item{n_fixed}{Number of fixed effects.}

\item{n_random}{Number of random effects.}

\item{start_resid_var}{Starting values residual variance, default starting
value equals 1.}

\item{start_random_var}{Starting values variance of random effects,
default starting values equal 1.}

\item{start_fixed_coef}{Starting values fixed effect coefficients,
default starting values equal 1.}

\item{prior_n}{Weight starting value residual variance.}

\item{prior_j}{Weight starting values fixed effects coefficients and random
effects variance.}
}
\value{
If theta is a list, it returns theta as it was supplied, else
  a list with needed objects is returned.
}
\description{
try_theta is a function which creates the model parameters and objects
  related to them. If the input is not a list, this function creates a list
  which is also a "sema" class with all needed objects stored within.
}
