% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitiateUnitParameters.R
\name{try_theta_j}
\alias{try_theta_j}
\title{try_theta_j is a function which creates the unit parameters and objects
  related to them. If the input is not a list, this function creates a list
  which is also a "sema" class with all needed objects stored within.}
\usage{
try_theta_j(theta_j, n_fixed, n_random, ids)
}
\arguments{
\item{theta_j}{Either an empty object or an object with class list and sema.}

\item{n_fixed}{Number of fixed effects.}

\item{n_random}{Number of random effects.}

\item{ids}{The id label}
}
\value{
If theta_j is a list, it returns theta_j as it was supplied, else
  a list with needed objects is returned.
}
\description{
try_theta_j is a function which creates the unit parameters and objects
  related to them. If the input is not a list, this function creates a list
  which is also a "sema" class with all needed objects stored within.
}
