% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupportFunctions.R
\name{update_average}
\alias{update_average}
\title{In this file all the functions needed by the SEMA algorithm are listed.
  NOTE that we use observation to refer to the level 1 data and
  individual or unit to refer to the units at level 2.
  So, the observations are clustered within individuals,
  but other groupings are dealt with the equally.
  NOTE 2, these functions are all read in by the SEMA algorithm meaning
  that you do not need to worry about transponing certain vectors or
  matrices, it has been taken care of in the main function.
update_average is the online update of a sample mean}
\usage{
update_average(old, obs, n)
}
\arguments{
\item{old}{The old (current) mean.}

\item{obs}{The new observation.}

\item{n}{The total number of observations.}
}
\value{
The updated sample mean.
}
\description{
In this file all the functions needed by the SEMA algorithm are listed.
  NOTE that we use observation to refer to the level 1 data and
  individual or unit to refer to the units at level 2.
  So, the observations are clustered within individuals,
  but other groupings are dealt with the equally.
  NOTE 2, these functions are all read in by the SEMA algorithm meaning
  that you do not need to worry about transponing certain vectors or
  matrices, it has been taken care of in the main function.
update_average is the online update of a sample mean
}
