% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupportFunctions.R
\name{update_id}
\alias{update_id}
\title{update_id is the online update of all unit level objects
  which do not depend on model parameters.}
\usage{
update_id(parameters_id, data_fixed, data_random, data_y)
}
\arguments{
\item{parameters_id}{A list of current state of the parameters,
@seealso \code{\link{build_theta_j}}.}

\item{data_fixed}{A vector with data of the fixed effects covariates.}

\item{data_random}{A vector with data of random effects covariates.}

\item{data_y}{A scalar with response or dependent variable.}
}
\value{
A list with updated objects of one unit.
}
\description{
update_id is the online update of all unit level objects
  which do not depend on model parameters.
}
