% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupportFunctions.R
\name{update_product}
\alias{update_product}
\title{update_product is the online update of a matrix product.
  NOTE: if you want to update the X'X matrix, part_a = x,
  part_b = t(x), where x = rx1 vector make sure the dimensions of the
  vectors of part_a and part_b match to compute the matrix product}
\usage{
update_product(old, part_a, part_b)
}
\arguments{
\item{old}{The old (current) matrix.}

\item{part_a}{A new observation, left or pre multiply part.}

\item{part_b}{A new observation, right or post multiply part.}
}
\value{
The updated matrix product.
}
\description{
update_product is the online update of a matrix product.
  NOTE: if you want to update the X'X matrix, part_a = x,
  part_b = t(x), where x = rx1 vector make sure the dimensions of the
  vectors of part_a and part_b match to compute the matrix product
}
