% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mstep.R
\name{update_x_inv}
\alias{update_x_inv}
\title{update_x_inv is the online update function of the inverse matrix.}
\usage{
update_x_inv(data_fixed, x_inv)
}
\arguments{
\item{data_fixed}{A fixed covariates vector from the new data point.}

\item{x_inv}{The current inverted matrix.}
}
\value{
The updated inverted matrix.
}
\description{
update_x_inv is the online update function of the inverse matrix.
}
