% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEMAFit.R
\name{sema_fit_one}
\alias{sema_fit_one}
\title{Fit multilevel models in a data stream I}
\usage{
sema_fit_one(data_fixed, data_random, data_y, id, theta_j, theta,
  print = FALSE, start_resid_var = 1, start_random_var = 1,
  start_fixed_coef = NULL, prior_n = 0, prior_j = 0)
}
\arguments{
\item{data_fixed}{A vector with the data of the fixed effects covariates.}

\item{data_random}{A vector with the data of the random effects covariates.}

\item{data_y}{A scalar with the response of this unit.}

\item{id}{A scalar which identifies the unit of this data point.}

\item{theta_j}{A list with this unit's parameters and contributions to the
sufficient statistics.}

\item{theta}{A list with model parameters and sufficient statistics.}

\item{print}{The default is FALSE, if TRUE the function
prints a summary of the model.}

\item{start_resid_var}{A scalar, optional if the user wants to provide a
start value of the residual variance, default start value is 1.}

\item{start_random_var}{A vector, optional if the user wants to provide a
start values of the variance of the random effects covariates, default
start value is 1. NOTE, if start values are provided make sure that the
length of the vector of start values matches the number of random effects.}

\item{start_fixed_coef}{A vector, optional if the user wants to provide
start values of the fixed effects, default is set to NULL such that
\code{sema_fit_one} creates the vector of start values matching the number
of fixed effects. NOTE, if start values are provided make sure that the
length of the vector of start values matches the number of fixed effects.}

\item{prior_n}{A scalar, if starting values are provided, prior_n determines
the weight of the starting value of the residual variance, default is 0.}

\item{prior_j}{A scalar, if starting values are provided, prior_j determines
the weight of the starting value of the variance of the random effects and
the fixed effects, default is 0.}
}
\value{
A list with a list with updated unit level parameters for one unit
  and a list with updated global parameters.
}
\description{
Fit a multilevel model online, row-by-row, without storing
  data points
}
\details{
This is the main function to fit the multilevel models in a data 
  stream. The function takes in one observation which consists of the id
  number of the unit the data of the fixed effects covariates, the values
  of the random effects covariates, the response or outcome, and the current
  state of the model parameters. Currently the algorithm can fit models including
  fixed effects at level 1 and 2 and random intercepts and slopes for
  continuous outcomes. The user manages storage and retrieval of unit's
  parameters. This function is also used in \code{\link{sema_fit_set}} and
  \code{\link{sema_fit_df}}.
}
\seealso{
\code{\link{sema_fit_set}}, \code{\link{sema_fit_df}},
\code{\link{summary_sema}}, \code{\link{sema_ranef}},
  \code{\link{sema_store_resid_var}}, \code{\link{sema_store_random_var}}, 
  \code{\link{sema_store_fixed_coef}}
}
\keyword{fitting}
\keyword{method}
\keyword{models}
\keyword{multilevel}
\keyword{online}
\keyword{stream}
