% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{store_fixed_coef}
\alias{store_fixed_coef}
\title{Store fixed effects coefficients}
\usage{
store_fixed_coef(object)
}
\arguments{
\item{object}{The sema output, a list containing the model parameters.}
}
\value{
A data frame with the number of observations and the fixed effects
  coefficients
}
\description{
This function extracts the fixed effects coefficients and saves
  them in a data frame with the number of data points seen so far.
}
\seealso{
\code{\link{ranef}}, \code{\link{store_random_var}}, 
  \code{\link{store_resid_var}}
}
\keyword{coefficients}
\keyword{save}
