% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_sema}
\alias{summary_sema}
\title{Interpreting sema output}
\usage{
summary_sema(x)
}
\arguments{
\item{x}{A sema model output.}
}
\value{
A list with sample size, number of units, the coefficients of the
  fixed effects, the variance of the random effects and the residual
  variance.
}
\description{
Returns a list with the current model parameter estimates
}
\details{
The output of the sema_fit functions are usually large, and 
  difficult to read, lists. In order to interpret the output of the sema_fit
  functions, \code{summary_sema} returns a general overview of the model 
  parameters, the fixed effects coefficients, the random effects variances
  and covariances, and the residual variance. The \code{summary_sema} 
  function is also incorpprated within the \code{sema_fit_one} function, 
  such that with the argument \code{print_every} the user can see a summary
  of the updated model parameters every X data points.
}
\seealso{
\code{\link{store_fixed_coef}}, 
   \code{\link{store_random_var}}, \code{\link{store_resid_var}}, 
   \code{\link{ranef}}
}
\keyword{model}
\keyword{sema}
\keyword{summary}
